Update playground to use CodeMirror editor
This adds syntax highlighting to the playground on the homepage and in
the tour.

It also shows errors where they happen in the editor. It's currently a
whole line, but with an update to the backend it will show an error
exactly where it happens (not just the line).

I have tried to keep the appearance and accessibility properties of the
textarea the same as much as possible. I think the outcome is pretty
neat: it looks just like any other Bootstrap input but with numbered
lines and syntax highlighting.

One major reason for choosing CodeMirror 6 is that it works well on
mobile.